server/requirements.txt
fastapi
pymongo
uvicorn
pydantic-core
